Conflicts:
	src/windows/VisualCppExpress2008/libengrid-VC2008/libengrid-VC2008.vcproj

+++ src/scripts/build-nglib.sh
@@ -4,7 +4,7 @@
 # +                                                                      +
 # + This file is part of enGrid.                                         +
 # +                                                                      +
-# + Copyright 2008-2010 enGits GmbH                                     +
+# + Copyright 2008-2012 enGits GmbH                                     +
 # +                                                                      +
 # + enGrid is free software: you can redistribute it and/or modify       +
 # + it under the terms of the GNU General Public License as published by +
@@ -25,12 +25,15 @@
 # DESCRIPTION:
 
 cd netgen_svn
-wget http://files.engits.eu/netgen-4.9.13-RC.tar.gz
-tar xzf netgen-4.9.13-RC.tar.gz
-rm netgen-4.9.13-RC.tar.gz
-rm -rf netgen-mesher
+wget http://engits.eu/files/netgen-4.9.13.zip
+unzip netgen-4.9.13.zip
+rm -f netgen-4.9.13.zip
+[ -e netgen-mesher ] && rm -rf netgen-mesher
 mkdir netgen-mesher
-mv netgen-4.9.13-RC netgen-mesher/netgen
+mv netgen-4.9.13 netgen-mesher/netgen
+cd netgen-mesher
+patch -p0 < ../nglib_engrid_mods.diff
+cd ..
 qmake
 make clean
 make -j4
